core protocol helpers for creating protocol evengs, services & threads
git-svn-id: http://svn.miranda-ng.org/main/trunk@5286 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/HTTPServer/src/IndexXML.cpp
@@ -116,7 +116,7 @@ bool bCreateIndexXML(const char * pszRealPath, const char * pszIndexPath,
 		strcpy(szBuffer, "index.xsl");
 	}
 
-	WriteFile(hFile, szBuffer, strlen(szBuffer), &dwBytesWritten, NULL);
+	WriteFile(hFile, szBuffer, (DWORD)strlen(szBuffer), &dwBytesWritten, NULL);
 
 	WriteFile(hFile, szXmlHeader2, sizeof(szXmlHeader2) - 1, &dwBytesWritten, NULL);
